Fixed styling, attr issues
There was a single space at the beginning of all lines that was throwing off my editors. Also added the ability to specify types for each attr that they will be auto-converted to upon opt collection. Added a quick check to see if the converted value was a number (this needs work to support any types), this works for my own immediate needs.